requirements.txt
# Specific repo for pytorch gpu
--extra-index-url https://download.pytorch.org/whl/cu130
torch
torchvision

# only for windows
pywin32==311 ; sys_platform == 'win32'

# base requirements
annotated-types==0.7.0
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
ipykernel==7.0.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
packaging==25.0
parso==0.8.5
pillow==11.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
setuptools==70.2.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
wcwidth==0.2.14

